+++ src/PayID/pay-id-components.ts
@@ -1,7 +1,29 @@
+// Disable multiple classes to accommodate the switch to idiomatic style naming.
+// TODO(keefertaylor): Remove this when migration is complete.
+/* eslint-disable  max-classes-per-file */
+
 /**
  * A class which encapsulates components of a Pay ID.
  */
-export default class PayIDComponents {
+export default class PayIdComponents {
+  /**
+   * Create a new PayIdComponents.
+   *
+   * @param host - The host of the payment pointer.
+   * @param path - The path of the payment pointer, starting with a leading '/'.
+   */
+  public constructor(
+    public readonly host: string,
+    public readonly path: string,
+  ) {}
+}
+
+/**
+ * A class which encapsulates components of a Pay ID.
+ *
+ * @deprecated Please use the idiomatically named `PayIdComponents` class instead.
+ */
+export class PayIDComponents {
   /**
    * Create a new PayIDComponents.
    *