+++ test/PayID/pay-id-utils.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai'
 
-import PayIDUtils from '../../src/PayID/pay-id-utils'
+import PayIdUtils from '../../src/PayID/pay-id-utils'
 import 'mocha'
 
 describe('PayIDUtils', function (): void {
@@ -11,7 +11,7 @@ describe('PayIDUtils', function (): void {
     const rawPayID = `${path}$${host}`
 
     // WHEN it is parsed to components.
-    const payIDComponents = PayIDUtils.parsePayID(rawPayID)
+    const payIDComponents = PayIdUtils.parsePayID(rawPayID)
 
     // THEN the host and path are set correctly.
     assert.equal(payIDComponents?.host, host)
@@ -25,7 +25,7 @@ describe('PayIDUtils', function (): void {
     const rawPayID = `${host}$${path}`
 
     // WHEN it is parsed to components.
-    const payIDComponents = PayIDUtils.parsePayID(rawPayID)
+    const payIDComponents = PayIdUtils.parsePayID(rawPayID)
 
     // THEN the Pay ID failed to parse.
     assert.isUndefined(payIDComponents)
@@ -38,7 +38,7 @@ describe('PayIDUtils', function (): void {
     const rawPayID = `${host}$${path}`
 
     // WHEN it is parsed to components.
-    const payIDComponents = PayIDUtils.parsePayID(rawPayID)
+    const payIDComponents = PayIdUtils.parsePayID(rawPayID)
 
     // THEN the Pay ID failed to parse.
     assert.isUndefined(payIDComponents)
@@ -51,7 +51,7 @@ describe('PayIDUtils', function (): void {
     const rawPayID = `${host}$${path}`
 
     // WHEN it is parsed to components.
-    const payIDComponents = PayIDUtils.parsePayID(rawPayID)
+    const payIDComponents = PayIdUtils.parsePayID(rawPayID)
 
     // THEN the Pay ID failed to parse.
     assert.isUndefined(payIDComponents)
@@ -62,6 +62,6 @@ describe('PayIDUtils', function (): void {
     const rawPayID = 'ZA̡͊͠͝LGΌIS̯͈͕̹̘̱ͮ$TO͇̹̺ͅƝ̴ȳ̳TH̘Ë͖́̉ ͠P̯͍̭O̚N̐Y̡'
 
     // WHEN it is parsed to components THEN the result is undefined
-    assert.isUndefined(PayIDUtils.parsePayID(rawPayID))
+    assert.isUndefined(PayIdUtils.parsePayID(rawPayID))
   })
 })